ld: put the binary at exact 4M
Shifting the starting address of a test case executable using SIZEOF_HEADERS
makes it non-trivial to load as a multiboot module. This patch ensures the entry
of every test case locates at exact 0x4000c (which is 0x40000 plus the size of
the multiboot header) by dropping the SIZEOF_HEADERS shift as well as build ids.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>